// apps/desktop/src/types/api/StockPosition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StockTransaction } from "./StockTransaction";

export type StockPosition = { symbol: string, quantity: number, averageCost: number, invested: number, marketPrice: number | null, marketValue: number | null, unrealisedProfitLoss: number | null, unrealisedPercent: number | null, realisedProfitLoss: number, transactions: Array<StockTransaction>, };